focus-voice-agent/requirements.txt
pipecat-ai[daily,speechmatics,google,silero]>=0.0.55
fastapi>=0.115
uvicorn>=0.34
httpx>=0.27
python-dotenv>=1.0
